garbage-collect-aws cli can iterate over aws resources and cli scans all ts files

+++ src/model/cloud-runner/cloud-runner-providers/aws/commands/aws-cli-commands.ts
@@ -7,23 +7,25 @@ export class AWSCLICommands {
   @CliFunction(`garbage-collect-aws`, `garbage collect aws`)
   static async garbageCollectAws() {
     process.env.AWS_REGION = Input.region;
+    CloudRunnerLogger.log(`Cloud Formation stacks`);
     const CF = new AWS.CloudFormation();
-
-    const stacks = (await CF.listStacks().promise()).StackSummaries?.filter(
-      (_x) => _x.StackStatus !== 'DELETE_COMPLETE',
-    );
+    const stacks =
+      (await CF.listStacks().promise()).StackSummaries?.filter((_x) => _x.StackStatus !== 'DELETE_COMPLETE') || [];
+    for (const element of stacks) {
+      CloudRunnerLogger.log(JSON.stringify(element, undefined, 4));
+      if (Input.cloudRunnerTests) await CF.deleteStack({ StackName: element.StackName }).promise();
+    }
+    CloudRunnerLogger.log(`ECS Clusters`);
+    const ecs = new AWS.ECS();
+    const clusters = (await ecs.listClusters().promise()).clusterArns || [];
     if (stacks === undefined) {
       return;
     }
-    CloudRunnerLogger.log(`Cloud Formation stacks`);
-    for (const element of stacks) {
-      CloudRunnerLogger.log(JSON.stringify(element, undefined, 4));
-      await CF.deleteStack({ StackName: element.StackName }).promise();
+    for (const element of clusters) {
+      const input: AWS.ECS.ListTasksRequest = {
+        cluster: element,
+      };
+      CloudRunnerLogger.log(JSON.stringify(await ecs.listTasks(input).promise(), undefined, 4));
     }
-
-    CloudRunnerLogger.log(`ECS Clusters`);
-    const ecs = new AWS.ECS();
-    CloudRunnerLogger.log(JSON.stringify(await ecs.listClusters().promise(), undefined, 4));
-    CloudRunnerLogger.log(JSON.stringify(await ecs.describeClusters().promise(), undefined, 4));
   }
 }
